requirements.txt
anthropic==0.86.0
APScheduler==3.10.4
Flask==3.1.3
gunicorn==23.0.0
openpyxl==3.1.5
supabase==2.28.3
google-genai>=0.8.0
PyMuPDF>=1.24.0
Pillow>=10.0.0
reportlab>=4.0.0
httpx>=0.27.0
